Moving integration tests from unit test folder to integration tests folder
