+++ common/src/it/resources/logback.xml
